+++ ErsatzTV.Infrastructure/Health/Checks/VaapiDriverHealthCheck.cs
@@ -1,45 +1,47 @@
-﻿using System;
-using System.Reflection;
+﻿using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using ErsatzTV.Core.Domain;
 using ErsatzTV.Core.FFmpeg;
 using ErsatzTV.Core.Health;
 using ErsatzTV.Core.Health.Checks;
-using ErsatzTV.Core.Interfaces.Repositories;
-using LanguageExt;
+using ErsatzTV.Infrastructure.Data;
+using Microsoft.EntityFrameworkCore;
 
 namespace ErsatzTV.Infrastructure.Health.Checks
 {
     public class VaapiDriverHealthCheck : BaseHealthCheck, IVaapiDriverHealthCheck
     {
-        private readonly IConfigElementRepository _configElementRepository;
+        private readonly IDbContextFactory<TvContext> _dbContextFactory;
 
-        public VaapiDriverHealthCheck(IConfigElementRepository configElementRepository) =>
-            _configElementRepository = configElementRepository;
+        public VaapiDriverHealthCheck(IDbContextFactory<TvContext> dbContextFactory)
+        {
+            _dbContextFactory = dbContextFactory;
+        }
 
         protected override string Title => "VAAPI Driver";
 
         public async Task<HealthCheckResult> Check()
         {
-            string version = Assembly.GetEntryAssembly()?.GetCustomAttribute<AssemblyInformationalVersionAttribute>()
-                ?.InformationalVersion ?? "unknown";
-
-            if (!version.Contains("docker", StringComparison.OrdinalIgnoreCase) ||
-                !version.Contains("vaapi", StringComparison.OrdinalIgnoreCase))
+            await using TvContext dbContext = _dbContextFactory.CreateDbContext();
+            List<FFmpegProfile> profiles = await dbContext.FFmpegProfiles
+                .Filter(p => p.HardwareAcceleration == HardwareAccelerationKind.Vaapi)
+                .ToListAsync();
+            
+            if (profiles.Count == 0)
             {
                 return NotApplicableResult();
             }
-            
-            Option<int> maybeVaapiDriver =
-                await _configElementRepository.GetValue<int>(ConfigElementKey.FFmpegVaapiDriver);
-            var vaapiDriver = (VaapiDriver)await maybeVaapiDriver.IfNoneAsync(0);
-            if (vaapiDriver == VaapiDriver.Default)
-            {
-                return InfoResult(
-                    "Settings > FFmpeg Settings > VAAPI Driver is set to Default; selecting iHD (Gen 8+) or i965 (up to Gen 9) may offer better performance");
-            }
 
-            return OkResult();
+
+            var defaultProfiles = profiles
+                .Filter(p => p.VaapiDriver == VaapiDriver.Default)
+                .ToList();
+
+            return defaultProfiles.Any()
+                ? InfoResult(
+                    $"{defaultProfiles.Count} FFmpeg Profile{(defaultProfiles.Count > 1 ? "s are" : " is")} set to use Default VAAPI Driver; selecting iHD (Gen 8+) or i965 (up to Gen 9) may offer better performance with Intel iGPU")
+                : OkResult();
         }
     }
 }